+++ packages/core-js-compat/src/data.mjs
@@ -1958,6 +1958,8 @@ export const data = {
   },
   'esnext.function.is-constructor': {
   },
+  'esnext.function.metadata': {
+  },
   // TODO: Remove from `core-js@4`
   'esnext.function.un-this': {
   },