+++ packages/core-js/actual/symbol/index.js
@@ -1,5 +1,7 @@
 var parent = require('../../stable/symbol');
 
+require('../../modules/esnext.function.metadata');
 require('../../modules/esnext.symbol.dispose');
+require('../../modules/esnext.symbol.metadata');
 
 module.exports = parent;