+++ packages/core-js/actual/symbol/metadata.js
@@ -1,0 +1,5 @@
+require('../../modules/esnext.function.metadata');
+require('../../modules/esnext.symbol.metadata');
+var WrappedWellKnownSymbolModule = require('../../internals/well-known-symbol-wrapped');
+
+module.exports = WrappedWellKnownSymbolModule.f('metadata');