+++ packages/core-js/full/function/metadata.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/function/metadata');
+
+module.exports = parent;