+++ packages/core-js/proposals/decorator-metadata-v2.js
@@ -1,2 +1,3 @@
 // https://github.com/tc39/proposal-decorator-metadata
+require('../modules/esnext.function.metadata');
 require('../modules/esnext.symbol.metadata');