+++ tests/unit-global/esnext.function.metadata.js
@@ -1,0 +1,4 @@
+QUnit.test('Function#@@metadata', assert => {
+  assert.true(Symbol.metadata in Function.prototype);
+  assert.same(Function.prototype[Symbol.metadata], null, 'is null');
+});