+++ tests/unit-pure/esnext.function.metadata.js
@@ -1,0 +1,6 @@
+import Symbol from 'core-js-pure/actual/symbol';
+
+QUnit.test('Function#@@metadata', assert => {
+  assert.true(Symbol.metadata in Function.prototype);
+  assert.same(Function.prototype[Symbol.metadata], null, 'is null');
+});